+++ zigzag_conversion.cpp
@@ -16,7 +16,7 @@ public:
   string convert(string s, int numRows) {
     if (numRows <= 1)
       return s;
-    const int len = (int)s.size();
+    const int len = s.size();
     string *str = new string[numRows];
     int row = 0, step = 1;
     for (int i = 0; i < len; ++i) {
